refactor: extracted out validateKeyAndCertificate and siblings into helpers
